Move context types to their own file
