+++ packages/server/src/__tests__/runQuery.test.ts
@@ -19,7 +19,7 @@ import type {
   GraphQLRequestListenerParsingDidEnd,
   GraphQLRequestListenerValidationDidEnd,
   BaseContext,
-} from '../types/index';
+} from '../externalTypes';
 import { InMemoryLRUCache } from 'apollo-server-caching';
 import type { ApolloServerOptions } from '../types';
 import { ApolloServer } from '../ApolloServer';
