+++ packages/server/src/externalTypes/context.ts
@@ -1,0 +1,4 @@
+export type BaseContext = {};
+
+export type ContextFunction<TContext extends BaseContext = BaseContext> =
+  () => Promise<TContext>;
