+++ packages/server/src/externalTypes/types.ts
@@ -65,8 +65,6 @@ export type HTTPGraphQLResponse = {
     }
 );
 
-export type BaseContext = {};
-
 export type WithRequired<T, K extends keyof T> = T & Required<Pick<T, K>>;
 
 /**
