+++ packages/server/src/runHttpQuery.ts
@@ -4,7 +4,7 @@ import type {
   BaseContext,
   HTTPGraphQLRequest,
   HTTPGraphQLResponse,
-} from './types/index';
+} from './externalTypes';
 import {
   ApolloServerInternals,
   internalExecuteOperation,
