+++ packages/server/src/utils/schemaInstrumentation.ts
@@ -9,7 +9,7 @@ import { defaultFieldResolver } from 'graphql/execution';
 import type {
   BaseContext,
   GraphQLRequestExecutionListener,
-} from '../types/index';
+} from '../externalTypes';
 
 export const symbolExecutionDispatcherWillResolveField = Symbol(
   'apolloServerExecutionDispatcherWillResolveField',
