+++ packages/server/src/utils/schemaManager.ts
@@ -3,7 +3,7 @@ import type {
   ApolloConfig,
   GraphQLExecutor,
   GraphQLSchemaContext,
-} from '../types/index';
+} from '../externalTypes';
 import type { Logger } from '@apollo/utils.logger';
 import type { GatewayInterface, Unsubscriber } from '../types';
 import type { SchemaDerivedData } from '../ApolloServer';
